+++ PSServicePrincipal/internal/functions/New-ServicePrincipal.ps1
@@ -102,7 +102,7 @@
             $securePassword = New-Object Microsoft.Azure.Commands.ActiveDirectory.PSADPasswordCredential -Property @{ StartDate = $currentDate; EndDate = $endCurrentDate; Password = $password}
             $newSpn = New-AzADServicePrincipal -ApplicationID $ApplicationID -PasswordCredential $securePassword -ErrorAction Stop
             $script:roleListToProcess.Add($newSpn)
-            $script:spnCounter ++
+            return
         }
     }
     catch
@@ -117,10 +117,12 @@
         {
             $password = Read-Host "Enter Password" -AsSecureString
             $securePassword = New-Object Microsoft.Azure.Commands.ActiveDirectory.PSADPasswordCredential -Property @{ StartDate = Get-Date; EndDate = Get-Date -Year 2024; Password = $password}
-            Write-PSFMessage -Level Host -Message "Creating SPN {0} - Secure Password {1}" -StringValues $newSPN.DisplayName, $newSPN.securePassword
             $newSPN = New-AzADServicePrincipal -DisplayName $DisplayName -PasswordCredential $securePassword -ErrorAction Stop
+            $BSTR = [System.Runtime.InteropServices.Marshal]::SecureStringToBSTR($password)
+            $UnsecureSecret = [System.Runtime.InteropServices.Marshal]::PtrToStringAuto($BSTR)
+            Write-PSFMessage -Level Host -Message "Creating SPN {0} - Secure Password {1}. WARNING: Backup this key!!! If you lose it you will need to reset the credentials for this SPN!" -StringValues $newSPN.DisplayName, $UnsecureSecret
             $script:roleListToProcess.Add($newSpn)
-            $script:spnCounter ++
+            return
         }
     }
     catch
@@ -137,10 +139,9 @@
             $newSpn = New-AzADServicePrincipal -DisplayName $DisplayName -ErrorAction Stop
             $BSTR = [System.Runtime.InteropServices.Marshal]::SecureStringToBSTR($newSpn.Secret)
             $UnsecureSecret = [System.Runtime.InteropServices.Marshal]::PtrToStringAuto($BSTR)
-            Write-PSFMessage -Level Host -Message "Creating SPN DisplayName {0} and secret {1}" -StringValues $DisplayName, $UnsecureSecret
-            Write-PSFMessage -Level Host -Message "WARNING: Backup this key!!! If you lose it you will need to reset the credentials for this SPN"
+            Write-PSFMessage -Level Host -Message "Creating SPN DisplayName {0} and secret {1}. WARNING: Backup this key!!! If you lose it you will need to reset the credentials for this SPN!" -StringValues $DisplayName, $UnsecureSecret
             $script:roleListToProcess.Add($newSpn)
-            $script:spnCounter ++
+            return
         }
     }
     catch
@@ -157,7 +158,7 @@
             Write-PSFMessage -Level Host -Message "Creating registered application and SPN {0}. Certificate uploaded to Azure application" -StringValues $DisplayName
             $newSPN = New-AzADServicePrincipal -DisplayName $DisplayName -CertValue $CertValue -StartDate $StartDate -EndDate $EndDate -ErrorAction Stop
             $script:roleListToProcess.Add($newSpn)
-            $script:appCounter ++
+            return
         }
     }
     catch